Actual parameters for RHEA Agent
